feat: export types from SplitPanel component for better type support
