+++ packages/vue-split-panel/src/index.ts
@@ -1,1 +1,2 @@
 export { default as SplitPanel } from './SplitPanel.vue';
+export * from './types';
